Changing floats to doubles and add Service if it doesn't already exist.

+++ GTFSDataFeeder/src/main/java/fi/ahto/example/traffic/data/gtfs/feeder/ShapeCollector.java
@@ -39,8 +39,8 @@ public class ShapeCollector {
         }
         
         ShapeData sd = new ShapeData();
-        sd.latitude = (float) sp.getLat();
-        sd.longitude = (float) sp.getLon();
+        sd.latitude = sp.getLat();
+        sd.longitude = sp.getLon();
         sd.seq = sp.getSequence();
         set.add(sd);
     }
